Split ckan
