+++ frictionless/plugins/ckan/__init__.py
@@ -1,0 +1,4 @@
+from .dialect import CkanDialect
+from .parser import CkanParser
+from .plugin import CkanPlugin
+from .storage import CkanStorage
